requirements.txt
asgiref==3.3.1
Django==3.1.6
django-autoslug==1.9.8
django-cors-headers==3.7.0
djangorestframework==3.12.2
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
